integrated CLGIndex into Anna object
