+++ anna.go
@@ -10,8 +10,9 @@ import (
 	"github.com/xh3b4sd/anna/factory/server"
 	"github.com/xh3b4sd/anna/gateway"
 	"github.com/xh3b4sd/anna/id"
-	// TODO
-	_ "github.com/xh3b4sd/anna/index/clg"
+	"github.com/xh3b4sd/anna/index/clg"
+	"github.com/xh3b4sd/anna/index/clg/collection"
+	"github.com/xh3b4sd/anna/index/clg/profile"
 	"github.com/xh3b4sd/anna/log"
 	"github.com/xh3b4sd/anna/net/char"
 	charexecnet "github.com/xh3b4sd/anna/net/char/net/exec"
@@ -80,6 +81,7 @@ func init() {
 }
 
 type annaConfig struct {
+	CLGIndex      spec.CLGIndex
 	CoreNet       spec.Network
 	FactoryServer spec.Factory
 	Log           spec.Log
@@ -88,7 +90,13 @@ type annaConfig struct {
 }
 
 func defaultAnnaConfig() annaConfig {
+	newCLGIndex, err := clg.NewIndex(clg.DefaultIndexConfig())
+	if err != nil {
+		panic(err)
+	}
+
 	newConfig := annaConfig{
+		CLGIndex:      newCLGIndex,
 		CoreNet:       nil,
 		FactoryServer: factoryserver.NewFactory(factoryserver.DefaultConfig()),
 		Log:           log.NewLog(log.DefaultConfig()),
@@ -101,7 +109,8 @@ func defaultAnnaConfig() annaConfig {
 
 func newAnna(config annaConfig) (spec.Anna, error) {
 	newAnna := &anna{
-		annaConfig:   config,
+		annaConfig: config,
+
 		BootOnce:     sync.Once{},
 		ID:           id.NewObjectID(id.Hex128),
 		Mutex:        sync.Mutex{},
@@ -109,14 +118,23 @@ func newAnna(config annaConfig) (spec.Anna, error) {
 		Type:         spec.ObjectType(objectTypeAnna),
 	}
 
-	newAnna.Log.Register(newAnna.GetType())
-
+	if newAnna.CLGIndex == nil {
+		return nil, maskAnyf(invalidConfigError, "CLG index must not be empty")
+	}
 	if newAnna.CoreNet == nil {
 		return nil, maskAnyf(invalidConfigError, "core network must not be empty")
+	}
+	if newAnna.Log == nil {
+		return nil, maskAnyf(invalidConfigError, "logger must not be empty")
 	}
 	if newAnna.Server == nil {
 		return nil, maskAnyf(invalidConfigError, "server must not be empty")
 	}
+	if newAnna.Storage == nil {
+		return nil, maskAnyf(invalidConfigError, "storage must not be empty")
+	}
+
+	newAnna.Log.Register(newAnna.GetType())
 
 	return newAnna, nil
 }
@@ -142,6 +160,9 @@ func (a *anna) Boot() {
 		go a.listenToSignal()
 		go a.writeStateInfo()
 
+		a.Log.WithTags(spec.Tags{L: "I", O: a, T: nil, V: 10}, "booting CLG index")
+		go a.CLGIndex.Boot()
+
 		a.Log.WithTags(spec.Tags{L: "I", O: a, T: nil, V: 10}, "booting factory")
 		go a.FactoryServer.Boot()
 
@@ -157,6 +178,7 @@ func (a *anna) Shutdown() {
 	a.Log.WithTags(spec.Tags{L: "D", O: a, T: nil, V: 13}, "call Shutdown")
 
 	a.ShutdownOnce.Do(func() {
+		go a.CLGIndex.Shutdown()
 		go a.CoreNet.Shutdown()
 		go a.FactoryServer.Shutdown()
 
@@ -380,6 +402,32 @@ func mainRun(cmd *cobra.Command, args []string) {
 	newCoreNet, err := corenet.NewCoreNet(newCoreNetConfig)
 	panicOnError(err)
 
+	// CLG collection
+	newCollectionConfig := collection.DefaultConfig()
+	newCollectionConfig.Log = newLog
+	newCollection, err := collection.New(newCollectionConfig)
+	panicOnError(err)
+
+	// CLG profile generator
+	newGeneratorConfig := profile.DefaultGeneratorConfig()
+	newGeneratorConfig.Collection = newCollection
+	newGeneratorConfig.Log = newLog
+	newGeneratorConfig.Storage = newStorage
+	newGenerator, err := profile.NewGenerator(newGeneratorConfig)
+	panicOnError(err)
+
+	// CLG index
+	newCLGConfig := clg.DefaultIndexConfig()
+	newCLGConfig.Generator = newGenerator
+	newCLGConfig.Log = newLog
+	newCLGIndex, err := clg.NewIndex(newCLGConfig)
+	panicOnError(err)
+
+	// log control
+	newLogControlConfig := logcontrol.DefaultConfig()
+	newLogControlConfig.Log = newLog
+	newLogControl := logcontrol.NewLogControl(newLogControlConfig)
+
 	// text interface
 	newTextInterfaceConfig := textinterface.DefaultConfig()
 	newTextInterfaceConfig.Log = newLog
@@ -387,11 +435,6 @@ func mainRun(cmd *cobra.Command, args []string) {
 	newTextInterfaceConfig.TextGateway = newTextGateway
 	newTextInterface, err := textinterface.NewTextInterface(newTextInterfaceConfig)
 	panicOnError(err)
-
-	// log control
-	newLogControlConfig := logcontrol.DefaultConfig()
-	newLogControlConfig.Log = newLog
-	newLogControl := logcontrol.NewLogControl(newLogControlConfig)
 
 	// server
 	newServerConfig := server.DefaultConfig()
@@ -405,6 +448,7 @@ func mainRun(cmd *cobra.Command, args []string) {
 
 	// anna
 	newAnnaConfig := defaultAnnaConfig()
+	newAnnaConfig.CLGIndex = newCLGIndex
 	newAnnaConfig.CoreNet = newCoreNet
 	newAnnaConfig.FactoryServer = newFactoryServer
 	newAnnaConfig.Log = newLog
