+++ index/clg/profile/generator_common.go
@@ -2,7 +2,13 @@ package profile
 
 import (
 	"reflect"
+
+	"github.com/xh3b4sd/anna/key"
 )
+
+func (g *generator) key(f string, v ...interface{}) string {
+	return key.NewSysKey(g, f, v...)
+}
 
 func (g *generator) isMethodValue(v reflect.Value) bool {
 	if !v.IsValid() {
